sweet alert added for delete confirmation

+++ src/Pages/MyToys/MyToys.jsx
@@ -1,12 +1,13 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../../AuthProvider/AuthProvider';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const MyToys = () => {
     const { user } = useContext(AuthContext);
     const email = user?.email;
     const [toys, setToys] = useState([]);
-    
+
     useEffect(() => {
         fetch(`https://action-house-server.vercel.app/mytoys/${email}`)
             .then(res => res.json())
@@ -14,25 +15,43 @@ const MyToys = () => {
     }, [email])
 
     const handleDelete = (_id) => {
-        console.log('delete', _id);
-        fetch(`http://localhost:5000/deletetoy/${_id}`,{
-            method:"DELETE"
-        })
-        .then(res=>res.json())
-        .then(data=>{
-            console.log(data);
-            if(data.deletedCount>0){
-                alert("delete successfully")
-                const remaining = toys.filter(user=>user._id !== _id)
-                setToys(remaining);
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://action-house-server.vercel.app/deletetoy/${_id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        if (data.deletedCount > 0) {
+                            
+                            const remaining = toys.filter(user => user._id !== _id)
+                            setToys(remaining);
+                        }
+                    })
+
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
             }
         })
+
     }
 
     return (
         <div className='mx-10'>
             <h1 className='text-center text-2xl font-bold text-black my-5'>My Toys</h1>
-            
+
 
             <table className='w-full bg-white border border-gray-300 text-center'>
                 <thead>
@@ -59,9 +78,9 @@ const MyToys = () => {
                             <td className="py-2 px-4 bg-blue-100 border border-gray-500 ">{toy.availableQuantity}</td>
                             <td className="py-2 px-4 bg-blue-100 border border-gray-500 "><button className='btn btn-primary bg-blue-500 hover:bg-blue-800 border-0 text-black'>View Details</button></td>
                             <td className="py-2 px-4 bg-blue-100 border border-gray-500 ">
-                                <Link to={`/updatetoy/${toy._id}`}><button  className='btn btn-outline btn-info border-0 text-black'>Update</button></Link>
+                                <Link to={`/updatetoy/${toy._id}`}><button className='btn btn-outline btn-info border-0 text-black'>Update</button></Link>
                             </td>
-                            <td className="py-2 px-4 bg-blue-100 border border-gray-500 "><button  className='btn btn-outline btn-warning border-0'  onClick={() => handleDelete(toy._id)} >Delete</button></td>
+                            <td className="py-2 px-4 bg-blue-100 border border-gray-500 "><button className='btn btn-outline btn-warning border-0' onClick={() => handleDelete(toy._id)} >Delete</button></td>
                         </tr>
                     ))}
                 </tbody>
